Use existing properties for username/password, move IT test, minor cleanup
